Speed up Postion#current_line
This optimization will lazily build a list of beginning line offsets which
are then searched via Array#bsearch_index.  I did not exhaustively test
this but in looping on tiny_markdown and calling it 1000 times the before
time (MRI 2.7) was ~2.8s and after went down to ~2.3s.  This is remarkable
if you consider the sample file is only 52 lines long!  If I extend the size
of sample.md to be 1990 lines long the time goes from 12m38s to 1m59s!!!