Fix line calculation issues.
The idea of adding a bogus -1 entry to line offsets ended up
causing the bisect to somehow return a value which is longer
than the actual number of lines of the source.

The fix is to remove that and just add the 1 index via addition.

Current problem is I have been unable to write up a simplified
test case showing the original problem so I have no test.