Also don't push 0 onto line_offsets
In the prior commit we removed the bogus -1 entry and added 1 to the
bsearch_index result, but this was insufficient.  Also removing the 0
entry by adding offsets only at the end of the line seems to be the
complete fix.

A test that checks for an error that is not at the first character
confirms this, passes with this commit and on 1.1.0, but fails on 1.2.0.